refactor: create separate 'GameHistory' component, render only if user registered
